LEVEL VIEWER COMPLETED!
"Finals DONT EDIT">"for dist">run.bat will run the level viewer.

DO NOT CHANGE ANYTHING HERE. If you want to play around with the level
viewer, copy and paste the "for dist" folder on your local computer and
edit the files within.

levels>default.txt is the level which is read by the program, editing
this file will edit the level that loads.

inside default.txt you can change which pictures are used for which
tiles, rearrange the tile placements, and change the background picture.